# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VentureForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ideas/{id}/vote:
    post:
      operationId: voteOnIdea
      tags: [votes]
      summary: Upvote or downvote an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Comments
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VoteBody:
      type: object
      properties:
        vote:
          type: string
          enum: [up, down]
      required: [vote]
    VoteResult:
      type: object
      properties:
        upvotes:
          type: integer
        downvotes:
          type: integer
        score:
          type: integer
        userVote:
          type: string
          enum: [up, down]
          nullable: true
      required: [upvotes, downvotes, score]
...